+++ ui/apps/dashboard/src/components/Insights/InsightsSQLEditor/functions/types.ts
@@ -1,0 +1,4 @@
+export type FunctionDescriptor = {
+  name: string;
+  signature: string;
+};
